requirements.txt
PySide6
selenium
webdriver-manager
python-dotenv
qdarkstyle
pyyaml
pandas
cryptography
openpyxl
requests